Settings statement refactor
