+++ src/imp/types/index.ts
@@ -1,4 +1,3 @@
-export * from "./builder";
 export * from "./http";
 export * from "./status";
 export * from "./validator";
