+++ test/standalone/standalone.spec.ts
@@ -9,6 +9,7 @@ import { standaloneByIdTests } from "./id";
 import { standaloneIfFilterTests } from "./ifFilter";
 import { standaloneInAreaTests } from "./inArea";
 import { standaloneIntersectTests } from "./intersect";
+import { standaloneOutTests } from "./out";
 import { standalonePivotTests } from "./pivot";
 import { standalonePolygonTests } from "./polygon";
 import { standaloneRecurseTests } from "./recurse";
@@ -27,4 +28,5 @@ describe("Standalone statements", () => {
 	describe("Around", standaloneAroundTests);
 	describe("In area", standaloneInAreaTests);
 	describe("Pivot", standalonePivotTests);
+	describe("Out", standaloneOutTests);
 });
